+++ lib/animated_splash_plus.dart
@@ -1,4 +1,4 @@
 library animated_splash_plus;
 
-export 'splash_widget.dart';
-export 'splash_controller.dart';
+export 'src/splash_widget.dart';
+export 'src/splash_controller.dart';